网关微服务平台初始化-调整代码，明确定义所有表必须有主键且需要定义为id,方便逻辑删除自动赋值为id

+++ gateway/gateway-admin-server/src/main/java/cn/cleanarch/gw/gateway/admin/system/service/SysMenuService.java
@@ -32,7 +32,7 @@ public interface SysMenuService extends IService<SysMenu> {
      * @param id 菜单ID
      * @return 成功、失败
      */
-    R removeMenuById(Integer id);
+    R removeMenuById(Long id);
 
     /**
      * 更新菜单信息
